requirements.txt
anyio==3.5.0
appnope==0.1.2
asgiref==3.5.0
backcall==0.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.1.2
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
docopt==0.6.2
entrypoints==0.3
et-xmlfile==1.1.0
fastapi==0.75.1
fonttools==4.28.5
git-filter-repo==2.34.0
h11==0.13.0
httptools==0.4.0
idna==3.2
ipykernel==6.5.0
ipython==7.29.0
jedi==0.18.0
jupyter-client==7.0.6
jupyter-core==4.9.1
kiwisolver==1.3.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.1
numpy==1.22.0
openpyxl==3.0.9
packaging==21.3
pandas==1.3.5
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
pipreqs==0.4.11
prompt-toolkit==3.0.22
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pydantic==1.9.0
Pygments==2.10.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
requests==2.26.0
scipy==1.8.0
seaborn==0.11.2
sniffio==1.2.0
soupsieve==2.3.1
starlette==0.17.1
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.6
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.2
yarg==0.1.9
